requirements.txt
textual>=0.60.0
Pillow>=10.0.0
pytest>=8.0.0
pytest-anyio>=0.0.0
anyio>=4.0.0
